feat(Connections): 가로 라인 연결선 포장 객체 신규
